+++ deepparse/dataset_container/tools.py
@@ -13,20 +13,13 @@ def former_python_list(tags: str) -> List:
     Return:
         A list of the parsed tag set.
     """
-    return tags.replace("[", "").replace("]", "").replace("'", "").split(", ")
-
-
-def comma_separated_list_reformat(tags: str) -> List:
-    """
-    Function to parse a comma separated "list" of tag.
-
-    Args:
-        tags (str): A tag set string to parse.
-
-    Return:
-        A list of the parsed tag set.
-    """
-    return tags.split(", ")
+    # We remove the [ and ] of the list.
+    # Then, we split each element using a comma as separator.
+    # Finally, since some case the element are separated by a comma (e.g. element1,element2)
+    # or a comma and a whitespace (e.g. element1, element2), we strip the whitespace on all tags to
+    # remove the trailing whitespace when element are separated by a coma and a whitespace.
+    # To fix https://github.com/GRAAL-Research/deepparse/issues/124.
+    return [tag.strip() for tag in tags.replace("[", "").replace("]", "").replace("'", "").split(",")]
 
 
 def validate_column_names(column_names: List[str]) -> bool: